refactor: location util functions (#421)
* chore: getDistance 중복 함수 제거

* chore: useLocation 제거

* refacotr: location util functions
